+++ android-sdk-couriers/src/main/java/com/citibox/courier/sdk/RetrievalLauncher.kt
@@ -8,17 +8,23 @@ import com.citibox.courier.sdk.deeplink.DeepLinkRetrievalContract
 import com.citibox.courier.sdk.domain.RetrievalParams
 import com.citibox.courier.sdk.domain.RetrievalResult
 import com.citibox.courier.sdk.domain.TransactionError
+import com.citibox.courier.sdk.webview.WebViewRetrievalContract
 
 class RetrievalLauncher private constructor(
     private val onResult: (RetrievalResult) -> Unit
 ) {
 
     private lateinit var deepLinkLauncher: ActivityResultLauncher<RetrievalParams>
+    private lateinit var webViewLauncher: ActivityResultLauncher<RetrievalParams>
 
     constructor(fragment: Fragment, onResult: (RetrievalResult) -> Unit) :
             this(onResult) {
         deepLinkLauncher = fragment.registerForActivityResult(
             /* contract = */ DeepLinkRetrievalContract(),
+            /* callback = */ onResult
+        )
+        webViewLauncher = fragment.registerForActivityResult(
+            /* contract = */ WebViewRetrievalContract(),
             /* callback = */ onResult
         )
     }
@@ -29,11 +35,24 @@ class RetrievalLauncher private constructor(
             /* contract = */ DeepLinkRetrievalContract(),
             /* callback = */ onResult
         )
+        webViewLauncher = activity.registerForActivityResult(
+            /* contract = */ WebViewRetrievalContract(),
+            /* callback = */ onResult
+        )
     }
 
     fun launch(params: RetrievalParams) {
         runCatching {
             deepLinkLauncher.launch(params)
+        }.getOrElse {
+            Log.w("RetrievalLauncher", "Error launching deep link", it)
+            launchWebView(params)
+        }
+    }
+
+    private fun launchWebView(params: RetrievalParams) {
+        runCatching {
+            webViewLauncher.launch(params)
         }.getOrElse {
             Log.w("RetrievalLauncher", "Error launching web view", it)
             fail()