+++ android-sdk-couriers/src/main/java/com/citibox/courier/sdk/webview/usecase/GetRetrievalUrlUseCase.kt
@@ -4,16 +4,12 @@ import androidx.core.net.toUri
 import com.citibox.courier.sdk.BuildConfig
 import com.citibox.courier.sdk.webview.models.WebAppEnvironment
 
-class GetDeliveryUrlUseCase {
+class GetRetrievalUrlUseCase {
 
     operator fun invoke(
         environment: WebAppEnvironment,
         accessToken: String,
-        tracking: String,
-        phone: String,
-        phoneHashed: String,
-        dimensions: String,
-        bookingId: String
+        citiboxId: String,
     ): String {
         val (base, segment) = when (environment) {
             WebAppEnvironment.Production -> BuildConfig.WEBAPP_PRO_URL to
@@ -33,21 +29,13 @@ class GetDeliveryUrlUseCase {
             .buildUpon()
             .appendPath(segment)
             .appendQueryParameter(PARAM_ACCESS_TOKEN, accessToken)
-            .appendQueryParameter(PARAM_TRACKING, tracking)
-            .appendQueryParameter(PARAM_PHONE, phone)
-            .appendQueryParameter(PARAM_PHONE_HASHED, phoneHashed)
-            .appendQueryParameter(PARAM_DIMENSIONS, dimensions)
-            .appendQueryParameter(PARAM_BOOKING_ID, bookingId)
+            .appendQueryParameter(PARAM_CITIBOX_ID, citiboxId)
             .build()
             .toString()
     }
 
     companion object {
         private const val PARAM_ACCESS_TOKEN = "access_token"
-        private const val PARAM_TRACKING = "tracking"
-        private const val PARAM_PHONE = "recipient_phone"
-        private const val PARAM_PHONE_HASHED = "recipient_hash"
-        private const val PARAM_DIMENSIONS = "dimensions"
-        private const val PARAM_BOOKING_ID = "booking_id"
+        private const val PARAM_CITIBOX_ID = "citibox_id"
     }
 }