+++ app/src/main/java/com/citibox/courier/sdk/example/delivery/DeliveryViewModel.kt
@@ -1,9 +1,9 @@
-package com.citibox.courier.sdk.example
+package com.citibox.courier.sdk.example.delivery
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.citibox.courier.sdk.example.models.MainSideEffect
-import com.citibox.courier.sdk.example.models.MainState
+import com.citibox.courier.sdk.example.delivery.models.DeliverySideEffect
+import com.citibox.courier.sdk.example.delivery.models.DeliveryState
 import com.citibox.courier.sdk.domain.DeliveryParams
 import com.citibox.courier.sdk.domain.DeliveryResult
 import com.citibox.courier.sdk.webview.models.WebAppEnvironment
@@ -14,12 +14,12 @@ import kotlinx.coroutines.flow.receiveAsFlow
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 
-class MainViewModel : ViewModel() {
+class DeliveryViewModel : ViewModel() {
 
-    private val _state = MutableStateFlow(MainState())
+    private val _state = MutableStateFlow(DeliveryState())
     val state = _state.asStateFlow()
 
-    private val _sideEffect = Channel<MainSideEffect>(Channel.BUFFERED)
+    private val _sideEffect = Channel<DeliverySideEffect>(Channel.BUFFERED)
     val sideEffect = _sideEffect.receiveAsFlow()
 
     fun onTokenChanged(value: String) = _state.update { it.copy(token = value) }
@@ -27,6 +27,8 @@ class MainViewModel : ViewModel() {
     fun onTrackingChanged(value: String) = _state.update { it.copy(tracking = value) }
 
     fun onPhoneChanged(value: String) = _state.update { it.copy(phone = value) }
+
+    fun onBookingIdChanged(value: String) = _state.update { it.copy(bookingId = value) }
 
     fun onPhoneHashedChanged(value: Boolean) = _state.update { it.copy(phoneHashed = value) }
 
@@ -40,18 +42,21 @@ class MainViewModel : ViewModel() {
 
             _state.update { it.copy(resultMessage = "") }
 
-            val param = DeliveryParams(
-                accessToken = _state.value.token,
-                tracking = _state.value.tracking,
-                recipientPhone = _state.value.phone,
-                isPhoneHashed = _state.value.phoneHashed,
-                dimensions = _state.value.dimensions,
-                webAppEnvironment = _state.value.environment
-            )
+            val param = _state.value.toDeliveryParams()
 
-            _sideEffect.trySend(MainSideEffect.Launch(param))
+            _sideEffect.trySend(DeliverySideEffect.Launch(param))
         }
     }
+
+    private fun DeliveryState.toDeliveryParams() = DeliveryParams(
+        accessToken = token,
+        tracking = tracking,
+        recipientPhone = phone,
+        isPhoneHashed = phoneHashed,
+        dimensions = dimensions,
+        bookingId = bookingId,
+        webAppEnvironment = environment
+    )
 
     fun onResultClear() {
         _state.update {