+++ app/src/main/java/com/citibox/courier/sdk/example/delivery/compose/EnvironmentSelector.kt
@@ -1,4 +1,4 @@
-package com.citibox.courier.sdk.example.compose
+package com.citibox.courier.sdk.example.delivery.compose
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement