+++ app/src/main/java/com/citibox/courier/sdk/example/delivery/models/DeliverySideEffect.kt
@@ -1,0 +1,9 @@
+package com.citibox.courier.sdk.example.delivery.models
+
+import com.citibox.courier.sdk.domain.DeliveryParams
+
+sealed class DeliverySideEffect {
+
+    data class Launch(val params: DeliveryParams) : DeliverySideEffect()
+
+}