+++ app/src/main/java/com/citibox/courier/sdk/example/delivery/models/DeliveryState.kt
@@ -1,13 +1,14 @@
-package com.citibox.courier.sdk.example.models
+package com.citibox.courier.sdk.example.delivery.models
 
 import com.citibox.courier.sdk.webview.models.WebAppEnvironment
 
-data class MainState(
+data class DeliveryState(
     val token: String = "",
     val tracking: String = "",
     val phone: String = "",
     val phoneHashed: Boolean = false,
     val dimensions: String = "",
+    val bookingId: String = "",
     val resultMessage: String = "",
     val environment: WebAppEnvironment = WebAppEnvironment.Sandbox
 )