+++ app/src/main/java/com/citibox/courier/sdk/example/retrieval/models/RetrievalSideEffect.kt
@@ -1,0 +1,9 @@
+package com.citibox.courier.sdk.example.retrieval.models
+
+import com.citibox.courier.sdk.domain.RetrievalParams
+
+sealed class RetrievalSideEffect {
+
+    data class Launch(val params: RetrievalParams): RetrievalSideEffect()
+
+}